Move remote background handling to ServicesManager. Use notification to detect loop completion. Move handleRemoteNotification to ServicesManager
