+++ Loop/Managers/LoopDataManager.swift
@@ -905,8 +905,6 @@ extension LoopDataManager {
         }
 
         updateRemoteRecommendation()
-        
-        self.delegate?.loopDataManager(self, loopDidFinishWithDosingDecision: dosingDecision, error: error)
     }
 
     fileprivate enum UpdateReason: String {
@@ -2208,14 +2206,6 @@ protocol LoopDataManagerDelegate: AnyObject {
     ///   - units: The recommended bolus in U
     /// - Returns: a supported bolus volume in U. The volume returned should be the nearest deliverable volume.
     func roundBolusVolume(units: Double) -> Double
-    
-    /// Informs the delegate that a Loop finished
-    ///
-    /// - Parameters:
-    ///   - manager: The manager
-    ///   - dosingDecision: The resulting dosing decision
-    ///   - error: An error, if any
-    func loopDataManager(_ manager: LoopDataManager, loopDidFinishWithDosingDecision: StoredDosingDecision, error: LoopError?)
 
     /// The pump manager status, if one exists.
     var pumpManagerStatus: PumpManagerStatus? { get }
